requirements.txt
torch torchvision
numpy
pandas
matplotlib
Pillow